\name{print.taug}
\alias{print.taug}

\title{
Printing an object of class taug
}
\description{
Prints an object of class taug.
}
\usage{
## S3 method for class 'taug'
\method{print}{taug}(x, ...)
}

\arguments{
  \item{x}{object of class taug.}
  \item{\dots}{further arguments to be passed to or from methods.}
}

\author{
Klaus Nordhausen
}

\keyword{multivariate}
\keyword{array}
