\name{print.tladle}
\alias{print.tladle}

\title{
Printing an object of class tladle
}
\description{
Prints an object of class tladle.
}
\usage{
## S3 method for class 'tladle'
\method{print}{tladle}(x, ...)
}

\arguments{
  \item{x}{object of class tladle.}
  \item{\dots}{further arguments to be passed to or from methods.}
}

\author{
Klaus Nordhausen
}

\keyword{multivariate}
\keyword{array}
