\name{tensorBSS-package}
\alias{tensorBSS-package}
\alias{tensorBSS}
\docType{package}
\title{
Blind Source Separation Methods for Tensor-Valued Observations
}
\description{
Contains several utility functions for manipulating tensor-valued data (centering, multiplication from a single mode etc.) and 
the implementations of the following blind source separation methods for tensor-valued data: \sQuote{tPCA}, \sQuote{tFOBI}, \sQuote{tJADE}, \sQuote{tgFOBI}, \sQuote{tgJADE} and \sQuote{tSOBI}.
}
\details{
\tabular{ll}{
Package: \tab tensorBSS\cr
Type: \tab Package\cr
Version: \tab 0.3.1\cr
Date: \tab 2017-08-22\cr
License: \tab GPL (>= 2)\cr
}


}
\author{
Joni Virta, Bing Li, Klaus Nordhausen and Hannu Oja 

Maintainer: Joni Virta <jomivi@utu.fi>
}

\references{
\cite{Virta, J., Li, B., Nordhausen, K. and Oja, H., (2016), Independent component analysis for tensor-valued data, \emph{submitted}, \bold{???},  ???--???. Preprint available on ArXiv {http://arxiv.org/abs/1602.00879}.} 

\cite{Virta, J., Li, B., Nordhausen, K. and Oja, H., (2016), JADE for Tensor-Valued Observation, \emph{submitted}, \bold{???},  ???--???. Preprint available on ArXiv { http://arxiv.org/abs/1603.05406}.}

\cite{Virta, J. and Nordhausen, K., (2016), Blind source separation of tensor-valued time series, \emph{submitted}, \bold{???},  ???--???.} 
}
\keyword{ package }
\keyword{multivariate}
\keyword{ts}
\keyword{array}


