\name{tensorBSS-package}
\alias{tensorBSS-package}
\alias{tensorBSS}
\docType{package}
\title{
Blind Source Separation Methods for Tensor-Valued Observations
}
\description{
Contains several utility functions for manipulating tensor-valued data (centering, multiplication from a single mode etc.) and 
the implementations of the following blind source separation methods for tensor-valued data: \sQuote{tPCA}, \sQuote{tFOBI}, \sQuote{tJADE}, \sQuote{k-tJADE}, \sQuote{tgFOBI}, \sQuote{tgJADE}, \sQuote{tSOBI}, \sQuote{tNSS.SD}, \sQuote{tNSS.JD} and \sQuote{tNSS.TD.JD}.
}
\details{
\tabular{ll}{
Package: \tab tensorBSS\cr
Type: \tab Package\cr
Version: \tab 0.3.3\cr
Date: \tab 2017-10-06\cr
License: \tab GPL (>= 2)\cr
}


}
\author{
Joni Virta, Bing Li, Klaus Nordhausen and Hannu Oja 

Maintainer: Joni Virta <jomivi@utu.fi>
}

\references{
\cite{Virta, J., Li, B., Nordhausen, K. and Oja, H., (2017), Independent component analysis for tensor-valued data, Journal of Multivariate Analysis, \doi{10.1016/j.jmva.2017.09.008}}

\cite{Virta, J., Li, B., Nordhausen, K. and Oja, H., (2016), JADE for Tensor-Valued Observation, \emph{submitted}, \bold{???},  ???--???. Preprint available on ArXiv { http://arxiv.org/abs/1603.05406}.}

\cite{Virta, J. and Nordhausen, K., (2017), Blind source separation of tensor-valued time series. Signal Processing 141, 204-216}  
}
\keyword{ package }
\keyword{multivariate}
\keyword{ts}
\keyword{array}


